requirements.txt
Pybooru==4.2.2
telegraph==2.1.0
Pyrogram==2.0.35
hentai==3.2.9
luscious==1.1.4
multporn==1.3.6
rule34==1.8.1
pygelbooru==0.3.2
Pillow>=9.0.0
python-dotenv>=0.19.1
TgCrypto==1.2.3
urllib3>=1.26.8